+++ src/collective/revisionmanager/tests/test_setup.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 """Setup tests for this package."""
-from collective.revisionmanager.testing import COLLECTIVE_REVISIONMANAGER_INTEGRATION_TESTING  # noqa
+from collective.revisionmanager.interfaces import IHistoryStatsCache
+from collective.revisionmanager.testing import COLLECTIVE_REVISIONMANAGER_INTEGRATION_TESTING  # noqa: E501
 from plone import api
+from plone.browserlayer.utils import registered_layers
 
 import unittest
 
@@ -22,12 +24,12 @@ class TestSetup(unittest.TestCase):
             'collective.revisionmanager'))
 
     def test_browserlayer(self):
-        """Test that ICollectiveRevisionmanagerLayer is registered."""
-        from collective.revisionmanager.interfaces import (
-            ICollectiveRevisionmanagerLayer)
-        from plone.browserlayer import utils
-        self.assertIn(ICollectiveRevisionmanagerLayer,
-                      utils.registered_layers())
+        layers = [l.getName() for l in registered_layers()]
+        self.assertIn('ICollectiveRevisionmanagerLayer', layers)
+
+    def test_persistent_utility(self):
+        sm = self.portal.getSiteManager()
+        self.assertIsNotNone(sm.getUtility(IHistoryStatsCache))
 
 
 class TestUninstall(unittest.TestCase):
@@ -43,3 +45,13 @@ class TestUninstall(unittest.TestCase):
         """Test if collective.revisionmanager is cleanly uninstalled."""
         self.assertFalse(self.installer.isProductInstalled(
             'collective.revisionmanager'))
+
+    def test_addon_layer_removed(self):
+        layers = [l.getName() for l in registered_layers()]
+        self.assertNotIn('ICollectiveRevisionmanagerLayer', layers)
+
+    def test_persistent_utility_removed(self):
+        from zope.component.interfaces import ComponentLookupError
+        with self.assertRaises(ComponentLookupError):
+            sm = self.portal.getSiteManager()
+            sm.getUtility(IHistoryStatsCache)